+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file. This change
 
 ## [Unreleased]
 
+- Add support for simple var nodes
+
 ## 0.1.3 (2019-11-27)
 
 - Basic implementation